Update Summary
Final draft of macro that will open a spreadsheet that contains the
adjusted variables. The program will save a text file version that will
be used as an input to the ACRU menu parameterization program. This
macro requires the user to input the directory that contains all
required file such as the MENU file, MENU_PARAM text file and the
fortran program to be in the same directory. The fortran program also
requires the same input files to be in the same directory for it to
work. Output files include: MENU_OLD, ORIGINAL_MENU,
ORIGINAL_MENU_PARAM, LOGRUN files as well as the log file for the macro
run. Revised August 11, 2015.